% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.R
\name{exit}
\alias{exit}
\title{Wrap-up function of R's native quit() function.}
\usage{
exit()
}
\description{
Quit the current working session without saving the current workspace.
}
\examples{
\dontrun{
exit()
}
}
